Compile time tweak for libcall lookup
If we have a large module which is mostly intrinsics, we hammer the lib call lookup path from CodeGenPrepare.  Adding a fastpath reduces compile by 15% for one such example.

The problem is really more general than intrinsics - a module with lots of non-intrinsics non-libcall calls has the same problem - but we might as well avoid an easy case quickly.



git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@360391 91177308-0d34-0410-b5e6-96231b3b80d8